Add files via upload
Here is the python code on which I was working through out the week. As of now the bot only has ability to reply back users with greeting, thanking and assist them to go to grader during office hours. Coming week I will work on code and add other abilities which a bot can do.